better cleaning of test output
